% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{addPersonalData}
\alias{addPersonalData}
\title{This function adds personal data to subjects: age, gender}
\usage{
addPersonalData(cohortData, connection, cdmSchema)
}
\arguments{
\item{cohortData}{Imported cohort data}

\item{connection}{Connection to database}

\item{cdmSchema}{Schema where "person" table resides}
}
\value{
A dataframe with selected patients. Columns: cohort_definition_id, subject_id, cohort_start_date, cohort_end_date
}
\description{
This function adds personal data to subjects: age, gender
}
\keyword{internal}
