% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbls.R
\name{check_tbl_coltypes}
\alias{check_tbl_coltypes}
\title{Check current & previous tables for shared column types.}
\usage{
check_tbl_coltypes(x)
}
\arguments{
\item{x}{a list containing tibble of the \code{previous} and \code{current} versions of
a table.}
}
\value{
a list containing the results of checks and associated metadata:
\itemize{
\item \code{check}: Check name
\item \code{valid}: Logical. Whether the table has any valid columns that can row
bound to compare.
\item \code{msg}: check result message
\item \code{valid_cols}: Names of columns that can be successfully row bound to be
compared.
\item \code{equal}: Logical vector. Whether columns share the same class.
\item \code{all_equal}: Logical. Whether all columns share the same class. If \code{TRUE} no
coercion of data types required.
}
}
\description{
Check current & previous tables for shared column types.
}
