% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbls.R
\name{check_tbls}
\alias{check_tbls}
\title{Check current & previous tables for shared variables and column types.}
\usage{
check_tbls(x)
}
\arguments{
\item{x}{a list containing tibble of the \code{previous} and \code{current} versions of
a table.}
}
\value{
a list containing the results of checks. \code{check_names} contains metadata
on column name checks and \code{check_coltypes} contains metadata on column data type
checks. The \code{valid} element states whether any columns in the two tables
can be compared.
}
\description{
Check current & previous tables for shared variables and column types.
}
