% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tbls.R
\name{combine_tbls}
\alias{combine_tbls}
\title{Combine current and previous tbl data}
\usage{
combine_tbls(current_tbl, previous_tbl)
}
\arguments{
\item{current_tbl}{a tibble of current data of a table.}

\item{previous_tbl}{a tibble of previous data of the same table.}
}
\value{
a combined (row bound) long table of current and previous data with
column \code{tbl} appended containing a factor indicator of whether a row originates
from the \code{current} or \code{previous} version of the data.
}
\description{
Combine current and previous tbl data
}
