% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_significantly_different_peptides.R
\name{diff_peptides_prep}
\alias{diff_peptides_prep}
\title{Peptides significantly different between states}
\usage{
diff_peptides_prep(df, replicates = 3, pv_cutoff = 0.01, alpha = 0.01)
}
\arguments{
\item{df}{differences in averages data.frame calculated using diff_ave function}

\item{replicates}{number of replicates}

\item{pv_cutoff}{p-value cutoff here set up to 0.01}

\item{alpha}{critical interval, here is multiple sets are using maximum CI is used.}
}
\value{
list of the dataframes
}
\description{
Returns list of the dataframes with significantly different peptides
}
