% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getChronologicalMatrix}
\alias{getChronologicalMatrix}
\title{Get chronological Markov state probabilities from raw data}
\usage{
getChronologicalMatrix(cohortData, stateCohorts)
}
\arguments{
\item{cohortData}{A data.frame object which is queried with getCohortsPatients function}

\item{stateCohorts}{The IDs of state cohorts}
}
\description{
This function outputs a data.frame object which describes the chronological movement of patients between the defined states
}
\keyword{internal}
