% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getInclusionTable}
\alias{getInclusionTable}
\title{Build table for cohorts which shows the inclusion statistics}
\usage{
getInclusionTable(cohortData, selectedCohorts, cohortLabels = NULL)
}
\arguments{
\item{cohortData}{Output of getCohortsPatients()}

\item{selectedCohorts}{Vector of included cohort ID's}

\item{cohortLabels}{Labels for cohort}
}
\value{
A dataframe with each selected cohort and its relation to the target cohort
}
\description{
This function creates an inclusion table from getCohortsPatients() function data
}
\keyword{internal}
