% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCHA-package.R
\docType{package}
\name{MATCHA-package}
\alias{MATCHA}
\alias{MATCHA-package}
\title{MATCHA: Data consistency report}
\description{
Load, compare and summarise IeDEA cohort data.
}
\author{
\strong{Maintainer}: Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID})

}
\keyword{internal}
