% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcha_ggplotly.R
\name{matcha_ggplotly}
\alias{matcha_ggplotly}
\title{Render a ggplot object as a plotly object with a caption and modified tooltip}
\usage{
matcha_ggplotly(p)
}
\arguments{
\item{p}{a ggplot plot object}
}
\value{
a ggplotly object with a caption and modified tooltip where appropriate
}
\description{
Render a ggplot object as a plotly object with a caption and modified tooltip
}
