% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{matcha_report_theme}
\alias{matcha_report_theme}
\title{MATCHA reports bslib theme}
\usage{
matcha_report_theme()
}
\value{
the MATCHA bslib theme adapted for reports
}
\description{
MATCHA reports bslib theme
}
