% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Create boxplot plot of numerical variable}
\usage{
plot_boxplot(
  tbl,
  x,
  include_violin = FALSE,
  include_points = FALSE,
  varwidth = TRUE,
  show_outliers = TRUE,
  log = FALSE
)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{x}{Character string. Name of variable in table to plot.}

\item{include_violin}{Logical. Include a violin plot?}

\item{include_points}{Logical. Include points in the plot?}

\item{varwidth}{Logical. If TRUE, width of boxplot is proportional to the
square root of the number of observations in each tbl.}

\item{show_outliers}{Logical. If TRUE, show outliers.}

\item{log}{Logical. If TRUE, log transform x variable with \code{log(x + 1)}
(to avoid returning \code{Inf} values) or \code{log(x + 1 -min(x))} if negative numbers
exist in \code{x} (to avoid returning \code{NaN} values).}
}
\value{
Histogram ggplot plot.
}
\description{
Create boxplot plot of numerical variable
}
