% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cat_count_by_year.R
\name{plot_cat_count_by_year}
\alias{plot_cat_count_by_year}
\title{Create faceted bar plots of category counts per year}
\usage{
plot_cat_count_by_year(tbl, date_col, y, mark_cutoff = TRUE, n = 4L)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{date_col}{Character string. Name of Date variable associated with \code{y}
from which year can be determined.}

\item{y}{Character string. Name of category variable in table to plot for which
counts per year will be plotted.}

\item{mark_cutoff}{Logical. Whether to show temporal cut-off of previous data.}

\item{n}{Integer. Number of most common categories to display. All other categories ignored.}
}
\value{
Stacked Bar plot ggplot plot.
}
\description{
Create faceted bar plots of category counts per year
}
