% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_by_date.R
\name{plot_count_by_date}
\alias{plot_count_by_date}
\title{Observation Frequency Timeline plot}
\usage{
plot_count_by_date(
  tbl,
  x,
  time_bin = c("day", "week", "month", "bimonth", "quarter", "season", "halfyear",
    "year"),
  position = c("stack", "dodge", "mirror", "diff"),
  plot_diff = FALSE,
  mark_cutoff = TRUE
)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{x}{Character string. Name of variable in table to plot.}

\item{time_bin}{Character string. Temporal bin.}

\item{position}{Character string.  Bar position.}

\item{plot_diff}{Logical. Whether to plot diff between previous and current.}

\item{mark_cutoff}{Logical. Whether to show temporal cut-off of previous data.}
}
\value{
Count by date ggplot plot.
(if \code{interactive} = TRUE).
}
\description{
Observation Frequency Timeline plot
}
