% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_events_by_year.R
\name{plot_events_by_year}
\alias{plot_events_by_year}
\title{Create nested bar plots of counts of two date variables per year.}
\usage{
plot_events_by_year(tbl, event1, event2, t0_tally = TRUE, mark_cutoff = TRUE)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{event1}{Character string. Name of Date variable containing records of occurrence of the first event.}

\item{event2}{Character string. Name of Date variable containing records of occurrence of the second event.}

\item{t0_tally}{Logical. Whether to count all records of \code{event2} in the year
\code{event1} is recorded or not.}

\item{mark_cutoff}{Logical. Whether to show temporal cut-off of previous data.}
}
\value{
Nested Dodge Bar plot ggplot plot.
}
\description{
Primarily used to compare \code{recart_d} to \code{enrol_d} but can be used to compare
the relative completion of two related events, completion of which is recorded
via dates each event occurred.
The first event, \code{event1}, is displayed as a transparent wider bar while
the second event, \code{event2}, which expected to always occur once the first
event is complete, is displayed as a solid, narrower bar within.
Note that if \code{t0_tally} is \code{TRUE}, all records of the second event (\code{event2}) are counted
in the year \code{event1} is recorded for a given record. Counting them in the same year
(regardless of when \code{event2} actually occured) allows for the correct display
of the proportion of \code{event1} records for which a record of \code{event2} also exists.
}
