% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Create bar plot of missing value (NA) counts across variables in a tbl}
\usage{
plot_missing(tbl, pk_tbl, exclude = NULL)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{pk_tbl}{Tibble of primary key information. Usually stored in userData$pk_tbl.}

\item{exclude}{Columns to exclude from plot.}
}
\value{
Missing value ggplot bar plot.
}
\description{
Create bar plot of missing value (NA) counts across variables in a tbl
}
