\name{rmc}
\alias{rmc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{R Mini Commander}
\description{
  This is a simple editor for R
}
\usage{
rmc()
}
%- maybe also 'usage' for other objects documented here.
%\details{
  
%}
\value{
  None
}
%\references{ ~put references to the literature/web site here ~ }
\author{Søren Højsgaard, sorenh@agrsci.dk}
\note{rmc requires the tcltk package to be installed}

%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }
\examples{
}
\keyword{utilities}% at least one, from doc/KEYWORDS

