% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_ci_timepoint.R
\name{sd_ci_timepoint}
\alias{sd_ci_timepoint}
\title{Calculate what would be the standard deviation based on the CI}
\usage{
sd_ci_timepoint(df, replicates = 3, alpha = 0.01)
}
\arguments{
\item{df}{input from tp function}

\item{replicates}{number of replicates in sample. Default set to 3.}

\item{alpha}{what alpha (CI cutoff should be used)}
}
\value{
data frame with standard deviation calculated from the CI
}
\description{
Calculate the CI to standard devaation, that can be used for the ellipsoid function
Calculate standard deviation from the CI using the equation as shown in:
SD=sqrt(replicates)*(uper_limit-lower_limit)/t
}
