% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Create histogram plot of numerical variable}
\usage{
plot_histogram(tbl, var, position = c("dodge", "mirror", "facet"), bins = 30L)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{var}{Character string. Name of variable in table to plot.}

\item{position}{Character string.  Bar position.}

\item{bins}{Integer. Number of bins.}
}
\value{
Histogram ggplot plot.
}
\description{
Create histogram plot of numerical variable
}
