% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Create histogram plot of numerical variable}
\usage{
plot_histogram(
  tbl,
  x,
  position = c("dodge", "mirror", "facet"),
  bins = 30L,
  log = FALSE
)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{x}{Character string. Name of variable in table to plot.}

\item{position}{Character string.  Bar position.}

\item{bins}{Integer. Number of bins.}

\item{log}{Logical. If TRUE, log transform x variable with \code{log(x + 1)}
(to avoid returning \code{Inf} values) or \code{log(x + 1 -min(x))} if negative numbers
exist in \code{x} (to avoid returning \code{NaN} values).}
}
\value{
Histogram ggplot plot.
}
\description{
Create histogram plot of numerical variable
}
