% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Missing Values plot}
\usage{
plot_missing(tbl, pk_tbl, exclude = NULL, compare_pk = FALSE)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{pk_tbl}{Tibble of primary key information. Usually stored in
\code{session$userData$pk[[session$userData$pk_tbl_name]]}.}

\item{exclude}{Columns to exclude from plot.}

\item{compare_pk}{Logical. Whether to compare table missing values to overall
primary key values.}
}
\value{
Missing value ggplot bar plot.
}
\description{
Missing Values plot
}
