% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_bar.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{Categorical Distribution plot}
\usage{
plot_stacked_bar(
  tbl,
  x,
  position = c("fill", "stack"),
  n = 8L,
  na.rm = TRUE,
  other.rm = FALSE
)
}
\arguments{
\item{tbl}{Combined table of previous and current data. Output of \code{\link[=combine_tbls]{combine_tbls()}}.}

\item{x}{Character string. Name of variable in table to plot.}

\item{position}{Character string.  Bar position. Either "fill" (percentages)
or "stack" (counts).}

\item{n}{Integer. Number of most common categories to display. All other
categories recoded to "Other".}

\item{na.rm}{Logical. Remove NA values from the table.}

\item{other.rm}{Logical. Remove "Other" values from the table.}
}
\value{
Stacked Bar plot ggplot plot.
}
\description{
Categorical Distribution plot
}
