% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_complete.R
\name{plot_summary_complete}
\alias{plot_summary_complete}
\title{Create bar plot of completeness of primary key ids across tables.}
\usage{
plot_summary_complete(pk, pk_col = "patient")
}
\arguments{
\item{pk}{List containing tibble of primary key information for each table loaded.
Usually stored in userData$pk_tbl.}

\item{pk_col}{Column name of primary key in each table. Default is "patient".}
}
\value{
Primary key completeness ggplot bar plot.
}
\description{
Create bar plot of completeness of primary key ids across tables.
}
